// src/api/models/sharingAgreementPartitionCoefficientResponseEndState.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

/**
 * How/why this coefficient's coverage ends
 */
export type SharingAgreementPartitionCoefficientResponseEndState = typeof SharingAgreementPartitionCoefficientResponseEndState[keyof typeof SharingAgreementPartitionCoefficientResponseEndState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SharingAgreementPartitionCoefficientResponseEndState = {
  OPEN: 'OPEN',
  OPEN_ORPHAN: 'OPEN_ORPHAN',
  PENDING_SUCCESSION: 'PENDING_SUCCESSION',
  DERIVED: 'DERIVED',
  CLOSED: 'CLOSED',
} as const;